优化代码

+++ veryrule-platform/src/main/java/com/lexinda/veryrule/platform/controller/VeryRuleElementController.java
@@ -21,11 +21,14 @@ import com.alibaba.fastjson2.JSONObject;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.lexinda.veryrule.annotation.VeryRuleSingle;
 import com.lexinda.veryrule.base.key.RuleCode;
+import com.lexinda.veryrule.bo.RuleBo;
 import com.lexinda.veryrule.platform.enums.VeryRuleElementGroupEnum;
 import com.lexinda.veryrule.platform.model.VeryRuleElementMenuModel;
 import com.lexinda.veryrule.platform.model.VeryRuleElementModel;
+import com.lexinda.veryrule.platform.model.VeryRuleFlowTempletModel;
 import com.lexinda.veryrule.platform.model.VeryRuleSceneModel;
 import com.lexinda.veryrule.platform.service.mybatis.VeryRuleElementMbService;
+import com.lexinda.veryrule.platform.service.mybatis.VeryRuleFlowTempletMbService;
 import com.lexinda.veryrule.platform.service.mybatis.VeryRuleSceneMbService;
 import com.lexinda.veryrule.vo.RestApiResponse;
 
@@ -47,6 +50,9 @@ public class VeryRuleElementController {
 	
 	@Autowired
 	private VeryRuleSceneMbService veryRuleSceneMbService;
+	
+	@Autowired
+	private VeryRuleFlowTempletMbService veryRuleFlowTempletMbService;
 
 	@RequestMapping(value = "/getVeryRuleElementList", method = RequestMethod.POST, produces = "application/json;charset=utf-8")
 	@VeryRuleSingle(ruleCode = RuleCode.NOTNULL, ruleKey = "currentPage,ruleType", ruleErrMsg = "不能为空")
@@ -129,6 +135,7 @@ public class VeryRuleElementController {
 							VeryRuleElementMenuModel.class);
 					if (StringUtils.isNotBlank(vgr.getGroupName())) {
 						veryRuleElementMenu.setRuleName(veryRuleElementMenu.getGroupName());
+						veryRuleElementMenu.setRuleCode(VeryRuleElementGroupEnum.getCode(vgr.getGroupName()));
 						if (!groupNameSet.contains(vgr.getGroupName())) {
 							groupNameSet.add(vgr.getGroupName());
 							veryRuleElementMenuList.add(veryRuleElementMenu);
@@ -221,6 +228,14 @@ public class VeryRuleElementController {
 			veryRuleElementData.setRuleCode(param.getString("ruleCode"));
 			veryRuleElementData.setRuleName(param.getString("ruleName"));
 			veryRuleElementData.setRuleType(param.getInteger("ruleType"));
+			String ruleExpr = param.getString("ruleExpr");
+			if(StringUtils.isNotBlank(ruleExpr)) {
+				veryRuleElementData.setRuleExpr(ruleExpr);
+			}
+			String ruleKey = param.getString("ruleKey");
+			if(StringUtils.isNotBlank(ruleKey)) {
+				veryRuleElementData.setRuleKey(ruleKey);
+			}
 			String ruleValue = param.getString("ruleValue");
 			if (StringUtils.isNotBlank(ruleValue)) {
 				veryRuleElementData.setRuleValue(ruleValue);
@@ -253,7 +268,7 @@ public class VeryRuleElementController {
 			veryRuleElementData.setVersion(1);
 			Map<String, Object> dataParam = new HashMap<String, Object>();
 			dataParam.put("ruleCode", veryRuleElementData.getRuleCode());
-			List<VeryRuleElementModel> veryRuleElementList = veryRuleElementService.selectVeryRuleElementList(param);
+			List<VeryRuleElementModel> veryRuleElementList = veryRuleElementService.selectVeryRuleElementList(dataParam);
 			if (veryRuleElementList.size() > 0) {
 				res.setErrorDesc("规则标识已存在");
 				res.setElapsedTime(System.currentTimeMillis() - res.getElapsedTime());
@@ -289,6 +304,14 @@ public class VeryRuleElementController {
 			dataParam.put("ruleDesc", param.getString("ruleDesc"));
 			dataParam.put("ruleType", param.getInteger("ruleType"));
 			dataParam.put("groupName", param.getString("groupName"));
+			String ruleExpr = param.getString("ruleExpr");
+			if(StringUtils.isNotBlank(ruleExpr)) {
+				dataParam.put("ruleExpr", ruleExpr);
+			}
+			String ruleKey = param.getString("ruleKey");
+			if(StringUtils.isNotBlank(ruleKey)) {
+				dataParam.put("ruleKey", ruleKey);
+			}
 			Map<String, Object> dataIdParam = new HashMap<String, Object>();
 			dataIdParam.put("id", id);
 			List<VeryRuleElementModel> veryRuleElement = veryRuleElementService.selectVeryRuleElementList(dataIdParam);
@@ -322,6 +345,41 @@ public class VeryRuleElementController {
 		try {
 			JSONObject param = JSON.parseObject(data);
 			Integer id = param.getInteger("id");
+			Map<String, Object> dataIdParam = new HashMap<String, Object>();
+			dataIdParam.put("id", id);
+			List<VeryRuleElementModel> veryRuleElement = veryRuleElementService.selectVeryRuleElementList(dataIdParam);
+			List<VeryRuleFlowTempletModel> veryRuleFlowTempletList = veryRuleFlowTempletMbService.selectVeryRuleFlowTempletList(new HashMap<String,Object>());
+			StringBuffer ruleTempStr = new StringBuffer();
+			veryRuleFlowTempletList.stream().forEach(rft->{
+				if(StringUtils.isNotBlank(rft.getRuleFlowTemplet())) {
+					try {
+						List<RuleBo> ruleList = JSON.parseArray(rft.getRuleFlowTemplet(), RuleBo.class);
+						for(RuleBo rl:ruleList) {
+							if(rl.getRuleCode().equals(veryRuleElement.get(0).getRuleCode())) {
+								ruleTempStr.append(rft.getRuleFlowTempletCode()).append(",");
+								break;
+							}
+						}
+					}catch(Exception e) {
+						JSONObject ruleData = JSON.parseObject(rft.getRuleFlowTemplet());
+						ruleData.entrySet().forEach(rd->{
+							if(!ruleTempStr.toString().contains(rft.getRuleFlowTempletCode())) {
+								List<RuleBo> ruleList = JSON.parseArray(rd.getValue().toString(), RuleBo.class);
+								for(RuleBo rl:ruleList) {
+									if(rl.getRuleCode().equals(veryRuleElement.get(0).getRuleCode())) {
+										ruleTempStr.append(rft.getRuleFlowTempletCode()).append(",");
+										break;
+									}
+								}
+							}
+						});
+					}
+				}
+			});
+			if(ruleTempStr.length()>0) {
+				res.setErrorDesc(ruleTempStr.substring(0,ruleTempStr.length()-1)+"正在使用此规则！");
+				return res;
+			}
 			veryRuleElementService.removeById(id);
 			res.setErrorCode(0);
 		} catch (Exception e) {
